Fix crash on Display Menu click.
* PVCombCol and PVCol are not staticaly distinct for now so they should
have the same behavior on INVALID_VALUE and overflow result. Otherwise,
Invalid result handling is not reached and crash could happen
everywhere.

+++ libinendi/src/PVAxesCombination.cpp
@@ -109,7 +109,7 @@ QString PVAxesCombination::to_string() const
 
 bool PVAxesCombination::is_last_axis(PVCombCol c) const
 {
-	return (c.value + 1) == _axes_comb.size();
+	return size_t(c.value + 1) == _axes_comb.size();
 }
 
 PVAxesCombination PVAxesCombination::serialize_read(PVCore::PVSerializeObject& so,
